fix: keep exists apt_proxy.conf.

+++ setup.sh
@@ -308,10 +308,12 @@ unset ftp_proxy
 unset all_proxy
 EOL
   if [ ! -z ${PROXY_URI} ];then
-    if [[ "${USE_PROXY}" == "ok" || "${USE_PROXY}" == "y" || "${USE_PROXY}" == "Y" || "${USE_PROXY}" == "yes" || "${USE_PROXY}" == "Yes"  || "${USE_PROXY}" == "YES" ]];then
-      echo "Acquire::http::Proxy \"${PROXY_URI}\";" >${SETUP_ACT_HOME}/.apt_proxy.conf
-    else
-      touch ${SETUP_ACT_HOME}/.apt_proxy.conf
+    if [[ ! -f ${SETUP_ACT_HOME}/.apt_proxy.conf ]];then
+      if [[ "${USE_PROXY}" == "ok" || "${USE_PROXY}" == "y" || "${USE_PROXY}" == "Y" || "${USE_PROXY}" == "yes" || "${USE_PROXY}" == "Yes"  || "${USE_PROXY}" == "YES" ]];then
+        echo "Acquire::http::Proxy \"${PROXY_URI}\";" >${SETUP_ACT_HOME}/.apt_proxy.conf
+      else
+        touch ${SETUP_ACT_HOME}/.apt_proxy.conf
+      fi
     fi
     cat >${SETUP_ACT_HOME}/.setpx.sh <<EOL
 #!/usr/bin/env bash
